test: add dedicated xattr and acl suites
